changes according to PR review
